Uso das validações no controller

+++ src/controllers/LojaController.ts
@@ -2,6 +2,7 @@ import {Request, Response} from 'express'
 import {LojaService} from '../services/lojaService'
 import {getDistancia} from '../api/ors'
 import { obterCoordenadas } from '../api/geocode';
+import { validateLojaStore, validateLojaUpdate, validateId } from '../validations/lojaValidation';
 
 
 class LojaController{
@@ -11,7 +12,13 @@ class LojaController{
         this.lojaService = new LojaService();
     }
 
-    async show(req: Request, res: Response){ //definir o tipo de retorno depois
+    async show(req: Request<{id: string}>, res: Response){ //definir o tipo de retorno depois
+        const errors = validateId(parseInt(req.params.id));
+
+        if (errors.length > 0) {
+          return res.status(400).json({ message: 'Erro de validação do ID', details: errors });
+        }
+
         const loja = await this.lojaService.findById(parseInt(req.params.id))
 
         res.status(200).json(loja);
@@ -24,18 +31,36 @@ class LojaController{
     }
 
     async store(req: Request, res: Response){
+        const errors = validateLojaStore(req.body);
+
+        if (errors.length > 0) {
+          return res.status(400).json({ message: 'Erro de validação', details: errors });
+        }
+
         const loja = await this.lojaService.create(req.body);
 
         res.status(201).json(loja);
     }
 
     async update(req: Request<{id: string}>, res: Response): Promise<Response>{
+        const errors = validateLojaUpdate(req.body);
+
+        if (errors.length > 0) {
+          return res.status(400).json({ message: 'Erro de validação', details: errors });
+        }
+
         const loja = await this.lojaService.update(parseInt(req.params.id), req.body)
 
         return res.status(200).json(loja);
     }
 
-    async destroy(req: Request, res: Response){
+    async destroy(req: Request<{id: string}>, res: Response){
+        const errors = validateId(parseInt(req.params.id));
+
+        if (errors.length > 0) {
+          return res.status(400).json({ message: 'Erro de validação do ID', details: errors });
+        }
+
         await this.lojaService.delete(parseInt(req.params.id));
 
         return res.json({message: 'Loja deletada com sucesso'})
@@ -55,6 +80,7 @@ class LojaController{
             })
         )).sort((lojaA, lojaB) => lojaA!.distancia - lojaB!.distancia);
         
+        return res.status(200).json(lojasProximas);
     }
 }
